refactored existing app styles into /app/styles
