+++ GoNOW/app/styles/DailyScreen.styles.ts
