+++ GoNOW/app/styles/Navigator.styles.tsx
